+++ src/virtual_dom/node/into_nodes.rs
@@ -14,8 +14,20 @@ impl<Ms> IntoNodes<Ms> for Node<Ms> {
     }
 }
 
+impl<Ms> IntoNodes<Ms> for Option<Node<Ms>> {
+    fn into_nodes(self) -> Vec<Node<Ms>> {
+        self.map(IntoNodes::into_nodes).unwrap_or_default()
+    }
+}
+
 impl<Ms> IntoNodes<Ms> for Vec<Node<Ms>> {
     fn into_nodes(self) -> Vec<Node<Ms>> {
         self
+    }
+}
+
+impl<Ms> IntoNodes<Ms> for Option<Vec<Node<Ms>>> {
+    fn into_nodes(self) -> Vec<Node<Ms>> {
+        self.unwrap_or_default()
     }
 }